Lazy module and rup estadisticas
